+++ test/bigtest/network/serializers/acquisitionMethod.js
@@ -1,3 +1,3 @@
 import { buildBaseSerializer } from '@folio/stripes-acq-components/test/bigtest/network';
 
-export default buildBaseSerializer('acquisitionMethods', 'acquisition_methods');
+export default buildBaseSerializer('acquisitionMethods', 'acquisitionMethods');